requirements.txt
gunicorn
whitenoise==6.4.0
environs
aiodns
aiogram
aiosignal
asgiref
async-timeout
attrs
Babel
Brotli
cchardet
certifi
cffi
charset-normalizer
config
decouple
Django
django-environ
djangorestframework
dj-database-url
frozenlist
idna
marshmallow
multidict
Pillow
psycopg2-binary
pycares
pycparser
python-decouple
python-dotenv
pytz
requests
sqlparse
telegram
thread6
tzdata
ujson
urllib3